don't use strum_macros directly, use strum with derive feature
The macros in the strum crate can be used either throught the strum derive
feature or directly through the strum_macros crate. The problem is that the
macro namespace is global and when two crates use strum differently (one through
strum_macros, one through the derive feature), the strum macros will conflict
and cause build errors.

In order to make badpod usable with other crates that use strum, this commit
switches over to strum with the derive feature as this is the recommended way by
the strum crate.

See also:
https://github.com/Peternator7/strum/issues/172
https://github.com/Peternator7/strum/issues/362